Rename folder in component
