+++ src/components/points-prelevement/point-exploitations.js
